pyproject.toml
[project]
name = "santander-cycles-pipeline"
version = "1.0.0"
description = "End-to-end batch data pipeline for TfL Santander Cycles usage data"
requires-python = ">=3.11"
dependencies = [
    "prefect>=2.14",
    "google-cloud-storage>=2.13",
    "google-cloud-bigquery>=3.13",
    "pyarrow>=14.0",
    "pandas>=2.1",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "dbt-bigquery>=1.7",
    "streamlit>=1.55.0",
    "plotly>=6.6.0",
    "watchdog>=6.0.0",
    "google-cloud-bigquery-storage>=2.36.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["scripts"]
